+++ src/views/note-details/note-details.tsx
@@ -17,6 +17,7 @@ import tocbot from "tocbot";
 import { PillButton } from "../../components/inputs/pill-button/pill-button";
 import { filterToColorMap } from "../notes/notes";
 import { relatedNotes } from "../../../pages/notes/[slug]";
+import { MDContent } from "../../components/display/md-content/md-content";
 
 interface IProps {
   noteDetails: INote;
@@ -62,51 +63,6 @@ export function NoteDetails({ noteDetails, relatedNotes }: Readonly<IProps>) {
     };
   }, [observedContentRef]);
 
-  const renderHeader = ({ id, children }) => {
-    return (
-      <div className={styles.mdHeader}>
-        <Link href={`#${id}`} scroll={false}>
-          <div className="js-toc-heading" id={id}>
-            <Text variant="h1">{children}</Text>
-          </div>
-        </Link>
-        <HorizontalLine />
-      </div>
-    );
-  };
-
-  const renderSubHeader = ({ children }) => {
-    return (
-      <div className={styles.mdSubHeader}>
-        <Text variant="h3">{children}</Text>
-      </div>
-    );
-  };
-
-  const renderImage = ({ children, src }) => {
-    return (
-      <div className={styles.codeBlock}>
-        <Image src={src} />
-      </div>
-    );
-  };
-
-  const renderUnorderedList = ({ children }) => {
-    return (
-      <ul>
-        <Text variant="p">{children}</Text>
-      </ul>
-    );
-  };
-
-  const renderParagraph = ({ children }) => {
-    return (
-      <div className={styles.mdParagraph}>
-        <Text variant="p">{children}</Text>
-      </div>
-    );
-  };
-
   const renderNoteHeader = () => {
     return (
       <header className={styles.header}>
@@ -129,55 +85,6 @@ export function NoteDetails({ noteDetails, relatedNotes }: Readonly<IProps>) {
     );
   };
 
-  const renderDetails = () => {
-    return (
-      <>
-        <div className="js-toc-content">
-          <Markdown
-            rehypePlugins={[rehypeSlug]}
-            components={{
-              h1: ({ id, children }) => renderHeader({ id, children }),
-              h2: ({ children }) => renderSubHeader({ children }),
-              p: ({ children }) => renderParagraph({ children }),
-              ul: ({ children }) => renderUnorderedList({ children }),
-              img: ({ children, src }) => renderImage({ children, src }),
-              a: ({ children, href }) => (
-                <TextLink
-                  href={href}
-                  color={filterToColorMap[noteDetails.category]}
-                  decoration="underline"
-                >
-                  {children}
-                </TextLink>
-              ),
-              code: ({ node, className, children, ...props }) => {
-                const match = /language-(\w+)/.exec(className || "");
-                return match ? (
-                  <div className={styles.codeBlock}>
-                    <SyntaxHighlighter
-                      style={atomDark}
-                      language={match[1]}
-                      PreTag="div"
-                      {...props}
-                    >
-                      {String(children).replace(/\n$/, "")}
-                    </SyntaxHighlighter>
-                  </div>
-                ) : (
-                  <span className={styles.inlineCode} {...props}>
-                    {children}
-                  </span>
-                );
-              },
-            }}
-          >
-            {noteDetails.content}
-          </Markdown>
-        </div>
-      </>
-    );
-  };
-
   return (
     <>
       <div className={styles.container}>
@@ -192,7 +99,7 @@ export function NoteDetails({ noteDetails, relatedNotes }: Readonly<IProps>) {
           >
             <NotesSideBar related={relatedNotes} current={noteDetails} />
           </div>
-          {renderDetails()}
+          <MDContent noteDetails={noteDetails} />
         </div>
       </div>
     </>
